# src/swagger/swaggerDocs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Document for ECOM API",
    "description": "Simple Node js Application with Auth, Product, Cart And Checkout\n\nMore Details Check [Swagger Editor](https://editor.swagger.io/)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/product/{id}": {
      "get": {
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Products"],
        "parameters": [
          {
            "name": "productID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "pen"
                  },
                  "description": {
                    "type": "string",
                    "example": "High Quality - Edit"
                  },
                  "price": {
                    "type": "integer",
                    "format": "double",
                    "example": 90.5
                  },
                  "MRP": {
                    "type": "integer",
                    "format": "double",
                    "example": 100
                  },
                  "profileImage": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Record deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "pen"
          },
          "description": {
            "type": "string",
            "example": "High Quality"
          },
          "price": {
            "type": "integer",
            "format": "double",
            "example": 90.5
          },
          "MRP": {
            "type": "integer",
            "format": "double",
            "example": 100
          },
          "productImage": {
            "type": "string",
            "example": "domain.tld/public/path/file.ext"
          },
          "createdAt": {
            "type": "string",
            "format": "datetime",
            "example": "2000-01-01T01:01:01.002Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "datetime",
            "example": "2000-01-02T01:01:01.002Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Custom Message"
          },
          "errorCodes": {
            "type": "string",
            "description": "Custom Error Code for Vernacular Language Compatibility "
          },
          "errors": {
            "type": "string",
            "description": "Actual Error Message"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}